+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 * Fixing `xan plot -T` granularity inference.
 * Fixing missing fractional seconds to default `datetime` serialization.
 * Fixing bin allocation with `xan bins --nice`.
+* Fixing `xan search --patterns -i`
 
 *Performance*